PresetInteractor: Introduced runtime presets
to provide better 3MF presets loading (for now loaded presets are always include as runtime ones)
+ introduced new API to get or traverse presets with respect to selected project (including the runtime presets)

+++ src/slic3r-shared/src/Slic3r/Biz/Config/FeatureValueJson.hpp
@@ -21,7 +21,7 @@ struct adl_serializer<Slic3r::Domain::Preset::FeatureValue>
         if (j.is_boolean())
             v = j.get<bool>();
         else if (j.is_number())
-            v = j.get<float>();
+            v = j.get<double>();
         else if (j.is_string())
             v = j.get<std::string>();
     }
